modify tests so they pass with pytest
