+++ src/gear/loaders.py
@@ -20,7 +20,7 @@ def readContract(contract):
     list_iocontracts = []
     for c in contract:
         if type(c) is not dict:
-            return ValueError("A dict type contract is expected.")
+            raise ValueError("A dict type contract is expected.")
         reqs = []
         for key in ["assumptions", "guarantees"]:
             reqs.append([polyhedralterm.PolyhedralTerm(term["coefficients"], term["constant"]) for term in c[key]])
@@ -37,7 +37,7 @@ def readContract(contract):
         return list_iocontracts
 
 
-def writeContract(contract, filename: str = None):
+def writeContract(contract, filename: str=None):
     """
     Converts a gear.iocontract.IoContract to a dictionary. If a list of iocontracts is passed,
     then a list of dicts is returned.
@@ -60,11 +60,11 @@ def writeContract(contract, filename: str = None):
         contract_dict["InputVars"] = [str(var) for var in c.inputvars]
         contract_dict["OutputVars"] = [str(var) for var in c.outputvars]
         contract_dict["assumptions"] = [
-            {"constant": str(term.constant), "coefficients": {str(k): str(v) for k, v in term.variables.items()}}
+            {"constant":term.constant, "coefficients": {str(k): v for k, v in term.variables.items()}}
             for term in c.a.terms
         ]
         contract_dict["guarantees"] = [
-            {"constant": str(term.constant), "coefficients": {str(k): str(v) for k, v in term.variables.items()}}
+            {"constant": term.constant, "coefficients": {str(k): v for k, v in term.variables.items()}}
             for term in c.g.terms
         ]
         contract_list.append(contract_dict)
