+++ tests/test_py_loaders.py
@@ -1,33 +1,13 @@
-import pytest
-from test_iocontract import validate_iocontract
-
 import gear.iocontract as iocontract
 from gear.loaders import readContract, writeContract
-
-
-def create_contracts(num=1):
-    """
-    Creates `num` number of contracts and returns a list of dicts
-    """
-    contracts = []
-    for i in range(num):
-        c = {
-            "InputVars": ["u" + str(i)],
-            "OutputVars": ["x" + str(i)],
-            "assumptions": [{"coefficients": {"u" + str(i): 1}, "constant": i}],
-            "guarantees": [{"coefficients": {"x" + str(i): 1}, "constant": 1}],
-        }
-        contracts.append(c)
-    if num == 1:
-        return contracts[0]
-    else:
-        return contracts
-
+from test_iocontract import validate_iocontract, create_contracts
+import pytest
 
 def test_read_contract():
+    # Create 1 contract and read it
     c = create_contracts(1)
-    io_c = readContract(c)
-    assert isinstance(io_c, iocontract.IoContract)
+    validate_iocontract(readContract(c))
+    # Create 5 contracts and read 
     c = create_contracts(5)
     io_contracts = readContract(c)
     assert len(io_contracts) == 5
@@ -36,12 +16,13 @@ def test_read_contract():
         assert validate_iocontract(io_c)
     # Ensure that all contracts are dictionaries
     c = [("InputVars", "u"), ("OutputVars", "x")]
-    with pytest.raises(ValueError, "A dict type contract is expected."):
+    with pytest.raises(ValueError, match="A dict type contract is expected."):
         readContract(c)
-
 
 def test_write_contract():
     c = create_contracts(1)
     io_c = readContract(c)
     assert c == writeContract(io_c)
     all_contracts = create_contracts(5)
+    io_contracts = readContract(all_contracts)
+    assert all_contracts == writeContract(io_contracts)
